requirements.txt
aiofiles==0.5.0
alembic==1.7.5
aniso8601==7.0.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
authy==2.2.6
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.20.13
botocore==1.23.13
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.7
distlib==0.3.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
fastapi-jwt-auth==0.5.0
fastapi-pagination==0.8.3
filelock==3.0.12
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
h11==0.12.0
httplib2==0.20.2
httptools==0.1.2
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
orjson==3.6.4
passlib==1.7.4
platformdirs==2.3.0
promise==2.3
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-http-client==3.3.3
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
s3transfer==0.5.0
selenium==3.141.0
sendgrid==6.7.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
starkbank-ecdsa==2.0.3
starlette==0.14.2
stripe==2.63.0
twilio==6.63.2
typing-extensions==3.10.0.2
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.16.0
virtualenv==20.7.2
watchgod==0.7
websockets==8.1
